+++ admin/app/favicon.ico
